Non-blocking context initialization. Fix "redeploy" cleanup.
Make sure that apps are started and become available in the order in which they are defined in the farm config.
Order was kind of already guaranted because gretty is using `[:]` for webapps list (which is LinkedHashMap), but previously web server was available only when all apps were initialized. So it was not possible to have apps to depend on the availability of other apps.

As a side-effect this change also fixes `redeployMode = "redeploy"` for jetty container. Previously old contexts were properly stopped only on first redeploy, but all subsequent redeploys didn't actually stop old contexts. Only new contexts were reachable from the web server but old ones were still hanging in the memory.

Signed-off-by: Artem Kozlov <artem@rem.sh>