Spring support: avoid classloading of webapp classes at configuration time
Previously, the Spring Boot main class was passed via a static field in
AppServletInitializer. This is cumbersome, because it requires to load
the class AppServletInitializer at configuration time, before the web
application has loaded.
In Jetty 10 the class loader is managed by Jetty, which creates the
web application class loader during the startup. Therefore the class loader
cannot be available at configuration time, leading to a NPE when using Spring
Boot.
As a replacement, the class name is now passed as init parameter of the servlet
context. This is a standard approach to passing initialization parameters during
servlet initialization, and avoids loading application classes at configuration
time.